Correct location of mattermost api client
